:hammer: refactor: Checkbox Component 리팩토링
