+++ src/components/Input/CheckBox.tsx
@@ -1,9 +1,12 @@
 import { forwardRef, InputHTMLAttributes, Ref } from 'react';
 import styled from 'styled-components';
-import { ReactComponent as Checked } from 'assets/icons/checked.svg';
-import { ReactComponent as UnChecked } from 'assets/icons/unchecked.svg';
+import { ReactComponent as CircleChecked } from 'assets/icons/checkbox_circle_checked.svg';
+import { ReactComponent as CircleUnChecked } from 'assets/icons/checkbox_circle_uhchecked.svg';
+import { ReactComponent as SquareUnChecked } from 'assets/icons/checkbox_square_unchecked.svg';
+import { ReactComponent as SquareChecked } from 'assets/icons/checkbox_square_checked.svg';
 
 export interface CheckboxProps extends InputHTMLAttributes<HTMLInputElement> {
+    shape: 'square' | 'circle';
     label?: string;
 }
 
@@ -19,7 +22,10 @@ const CheckBoxLabel = styled.label`
     align-items: center;
 `;
 
-const HiddenCheckbox = styled.input<CheckboxProps>`
+const HiddenCheckbox = styled.input.attrs<CheckboxProps>((props) => ({
+    type: 'checkbox',
+    shape: props.shape,
+}))`
     border: 0;
     clip: rect(0 0 0 0);
     height: 1px;
@@ -40,7 +46,13 @@ const StyledP = styled.p<{ checked: boolean }>`
 
 const Checkbox = forwardRef(
     (
-        { label, id, checked = false, ...props }: CheckboxProps,
+        {
+            shape = 'square',
+            label,
+            id,
+            checked = false,
+            ...props
+        }: CheckboxProps,
         ref: Ref<HTMLInputElement>,
     ) => {
         return (
@@ -52,7 +64,11 @@ const Checkbox = forwardRef(
                         ref={ref}
                         {...props}
                     />
-                    {checked ? <Checked /> : <UnChecked />}
+                    {shape === 'circle' && checked && <CircleChecked />}
+                    {shape === 'circle' && !checked && <CircleUnChecked />}
+
+                    {shape === 'square' && checked && <SquareChecked />}
+                    {shape === 'square' && !checked && <SquareUnChecked />}
                     <StyledP checked={checked}>{label}</StyledP>
                 </CheckBoxLabel>
             </CheckboxContainer>
