+++ src/components/Modal/CouponListModal.tsx
@@ -9,8 +9,8 @@ import { orderSheet } from 'api/order';
 import { CouponApplyResponse, CouponRequest } from 'models/order';
 import { CHANNEL_TYPE } from 'models';
 import { useTypedSelector } from 'state/reducers';
-import { ReactComponent as Checked } from 'assets/icons/checkBox_circle_checked.svg';
-import { ReactComponent as UnChecked } from 'assets/icons/checkBox_circle.svg';
+import { ReactComponent as Checked } from 'assets/icons/checkbox_circle_checked.svg';
+import { ReactComponent as UnChecked } from 'assets/icons/checkbox_circle_uhchecked.svg';
 import { ReactComponent as Logo } from 'assets/logo/headerLogo.svg';
 import CouponBackgroundWhite from 'assets/images/coupon_background_white.svg';
 import CouponBackgroundBlack from 'assets/images/coupon_background_black.svg';
